feat: cria job para mandar email de inatividade
