+++ package.json
@@ -8,7 +8,6 @@
     "build": "tsc",
     "start": "node dist/index.js",
     "dev": "ts-node src/index.ts",
-    "worker": "ts-node src/worker.ts",
     "prisma:generate": "prisma generate",
     "prestart": "npm run build",
     "format": "prettier --write .",
