+++ src/routes/job.routes.ts
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+
+import { jobController } from '../controllers/JobController';
+
+export const jobRoutes = Router();
+
+jobRoutes.post('/inatividade', async (req, res) => {
+  await jobController.runInatividadeJob(req, res);
+});
